requirements.txt
absl-py==2.1.0
actionlib==1.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alabaster==0.7.8
angles==1.9.13
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.5.2
apturl==0.5.2
argcomplete==1.8.1
astunparse==1.6.3
async-timeout==4.0.3
atomicwrites==1.1.5
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
Babel==2.6.0
base-local-planner==1.17.3
bcrypt==3.1.7
beautifulsoup4==4.8.2
blinker==1.8.2
bondpy==1.8.6
breezy==3.0.2
Brlapi==0.7.0
cachetools==5.5.0
camera-calibration==1.17.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
cbor==1.0.0
certifi==2019.11.28
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
click==8.1.7
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.2.1
colcon-cmake==0.2.29
colcon-common-extensions==0.3.0
colcon-core==0.18.4
colcon-defaults==0.2.8
colcon-devtools==0.3.0
colcon-installed-package-information==0.2.1
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-override-check==0.0.1
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.9
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colcon-zsh==0.5.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.1
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cov-core==1.15.0
coverage==4.5.2
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.16.2
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distlib==0.3.0
distro==1.9.0
distro-info==0.23+ubuntu1.1
docutils==0.16
dulwich==0.19.15
dynamic-reconfigure==1.7.3
empy==3.3.2
entrypoints==0.3
exceptiongroup==1.2.2
fastimport==0.9.8
filelock==3.16.1
flake8==3.7.9
flask==3.0.3
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.4.0
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
geodesy==0.5.6
google-auth==2.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpg==1.13.1
grpcio==1.67.0
h11==0.14.0
h5py==3.11.0
html5lib==1.0.1
httpcore==1.0.6
httplib2==0.14.0
httpx==0.27.2
hyperlink==19.0.0
idna==2.8
ifcfg==0.18
image-geometry==1.16.2
imagesize==1.2.0
importlib-metadata==8.5.0
importlib-resources==6.4.5
incremental==16.10.1
interactive-markers==1.12.0
itsdangerous==2.2.0
jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
keras==2.13.1
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
lark-parser==0.8.1
laser-geometry==1.6.7
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==18.1.1
llvmlite==0.41.1
louis==3.12.0
lxml==4.5.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mccabe==0.6.1
message-filters==1.17.0
mnist==0.2.2
mock==3.0.5
more-itertools==4.2.0
mpi4py==3.0.3
mpmath==1.3.0
multidict==6.1.0
mypy==0.761
mypy-extensions==0.4.3
netifaces==0.10.4
networkx==3.1
nicla-vision-ros==1.0.2
nose==1.3.7
nose2==0.9.1
notify2==0.3
numba==0.58.1
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.0
olefile==0.46
openai==0.28.0
openai-whisper==20240930
opt-einsum==3.4.0
packaging==20.3
pandas==2.0.3
paramiko==2.6.0
pbr==5.4.5
pexpect==4.6.0
Pillow==7.0.0
pluggy==0.13.0
propcache==0.2.0
protobuf==4.25.5
psutil==5.5.1
py==1.8.1
py-ubjson==0.14.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAudio==0.2.14
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.22
pycryptodomex==3.6.1
pycups==1.9.73
pydantic==2.9.2
pydantic-core==2.23.4
pydocstyle==2.1.1
pydot==1.4.1
pyflakes==2.1.1
PyGithub==1.43.7
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
pyproj==2.5.0
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
pyserial==3.5
pytest==4.6.9
pytest-cov==2.8.1
pytest-mock==1.10.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.1
python-gitlab==2.0.1
python-gnupg==0.4.5
python-qt-binding==0.4.4
python-snappy==0.5.3
PyTrie==0.2
pyttsx3==2.98
pytz==2024.2
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
regex==2024.9.11
reportlab==3.5.34
requests==2.32.3
requests-oauthlib==2.0.0
requests-unixsocket==0.2.0
resource-retriever==1.12.7
roman==2.0.0
rosbag==1.17.0
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosdep==0.25.1
rosdep-modules==0.25.1
rosdistro==1.0.1
rosdistro-modules==1.0.1
rosgraph==1.17.0
rosinstall==0.7.8
rosinstall-generator==0.1.23
roslaunch==1.17.0
roslib==1.15.8
roslint==0.12.0
roslz4==1.17.0
rosmake==1.15.8
rosmaster==1.17.0
rosmsg==1.17.0
rosnode==1.17.0
rosparam==1.17.0
rospkg==1.5.1
rospkg-modules==1.5.1
rospy==1.17.0
rosserial-python==0.9.2
rosservice==1.17.0
rostest==1.17.0
rostopic==1.17.0
rosunit==1.15.8
roswtf==1.17.0
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.12
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.17
rqt-launch==0.4.9
rqt-logger-level==0.4.12
rqt-moveit==0.5.11
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.4
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
rsa==4.9
rviz==1.14.25
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
SecretStorage==2.3.1
sensor-msgs==1.13.1
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
smach==2.5.2
smach-ros==2.5.2
smclib==1.8.6
sniffio==1.3.1
snowballstemmer==2.0.0
sounddevice==0.5.1
soupsieve==1.9.5
SpeechRecognition==3.10.4
Sphinx==1.8.5
sympy==1.13.3
systemd-python==234
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.7
tf2-kdl==0.7.7
tf2-py==0.7.7
tf2-ros==0.7.7
threadpoolctl==3.5.0
tiktoken==0.7.0
toml==0.10.0
topic-tools==1.17.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.6
triton==3.0.0
turtlebot3-example==1.2.5
turtlebot3-teleop==1.2.5
Twisted==18.9.0
txaio==2.10.0
typed-ast==1.4.1
typing-extensions==4.12.2
tzdata==2024.2
u-msgpack-python==2.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
unique-id==1.0.6
urllib3==1.25.8
vcstools==0.1.42
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
werkzeug==3.0.5
whisper==1.1.10
wrapt==1.11.2
wsaccel==0.6.2
wstool==0.1.18
xacro==1.14.18
xkit==0.0.0
yarl==1.15.2
youtube-dl==2020.3.24
zipp==3.20.2
zope.interface==4.7.1
